+++ torchmetrics/classification/hinge.py
@@ -83,6 +83,8 @@ class Hinge(Metric):
         >>> hinge(preds, target)
         tensor([2.2333, 1.5000, 1.2333])
     """
+    measure: Tensor
+    total: Tensor
 
     def __init__(
         self,
@@ -92,7 +94,7 @@ class Hinge(Metric):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
-    ):
+    ) -> None:
         super().__init__(
             compute_on_step=compute_on_step,
             dist_sync_on_step=dist_sync_on_step,
@@ -113,7 +115,7 @@ class Hinge(Metric):
         self.squared = squared
         self.multiclass_mode = multiclass_mode
 
-    def update(self, preds: Tensor, target: Tensor):
+    def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
         measure, total = _hinge_update(preds, target, squared=self.squared, multiclass_mode=self.multiclass_mode)
 
         self.measure = measure + self.measure