+++ torchmetrics/classification/roc.py
@@ -114,7 +114,7 @@ class ROC(Metric):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
-    ):
+    ) -> None:
         super().__init__(
             compute_on_step=compute_on_step,
             dist_sync_on_step=dist_sync_on_step,
@@ -133,7 +133,7 @@ class ROC(Metric):
             ' For large datasets this may lead to large memory footprint.'
         )
 
-    def update(self, preds: Tensor, target: Tensor):
+    def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
         """
         Update state with predictions and targets.
 
@@ -165,6 +165,8 @@ class ROC(Metric):
         """
         preds = torch.cat(self.preds, dim=0)
         target = torch.cat(self.target, dim=0)
+        if not self.num_classes:
+            raise ValueError(f'`num_classes` bas to be positive number, but got {self.num_classes}')
         return _roc_compute(preds, target, self.num_classes, self.pos_label)
 
     @property