+++ torchmetrics/image/ssim.py
@@ -63,7 +63,7 @@ class SSIM(Metric):
         compute_on_step: bool = True,
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
-    ):
+    ) -> None:
         super().__init__(
             compute_on_step=compute_on_step,
             dist_sync_on_step=dist_sync_on_step,
@@ -84,7 +84,7 @@ class SSIM(Metric):
         self.k2 = k2
         self.reduction = reduction
 
-    def update(self, preds: Tensor, target: Tensor):
+    def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
         """
         Update state with predictions and targets.
 
@@ -96,7 +96,7 @@ class SSIM(Metric):
         self.y_pred.append(preds)
         self.y.append(target)
 
-    def compute(self):
+    def compute(self) -> Tensor:
         """
         Computes explained variance over state.
         """