+++ torchmetrics/regression/cosine_similarity.py
@@ -68,7 +68,7 @@ class CosineSimilarity(Metric):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
-    ):
+    ) -> None:
         super().__init__(
             compute_on_step=compute_on_step,
             dist_sync_on_step=dist_sync_on_step,
@@ -80,7 +80,7 @@ class CosineSimilarity(Metric):
         self.add_state("target", [], dist_reduce_fx="cat")
         self.reduction = reduction
 
-    def update(self, preds: Tensor, target: Tensor):
+    def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
         """
         Update metric states with predictions and targets.
 
@@ -94,7 +94,7 @@ class CosineSimilarity(Metric):
         self.preds.append(preds)
         self.target.append(target)
 
-    def compute(self):
+    def compute(self) -> Tensor:
         preds = dim_zero_cat(self.preds)
         target = dim_zero_cat(self.target)
         return _cosine_similarity_compute(preds, target, self.reduction)