+++ torchmetrics/regression/mean_absolute_error.py
@@ -55,7 +55,7 @@ class MeanAbsoluteError(Metric):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
-    ):
+    ) -> None:
         super().__init__(
             compute_on_step=compute_on_step,
             dist_sync_on_step=dist_sync_on_step,
@@ -66,7 +66,7 @@ class MeanAbsoluteError(Metric):
         self.add_state("sum_abs_error", default=tensor(0.0), dist_reduce_fx="sum")
         self.add_state("total", default=tensor(0), dist_reduce_fx="sum")
 
-    def update(self, preds: Tensor, target: Tensor):
+    def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
         """
         Update state with predictions and targets.
 
@@ -79,7 +79,7 @@ class MeanAbsoluteError(Metric):
         self.sum_abs_error += sum_abs_error
         self.total += n_obs
 
-    def compute(self):
+    def compute(self) -> Tensor:
         """
         Computes mean absolute error over state.
         """